+++ lib/src/main/java/com/shopify/checkoutsheetkit/BaseWebView.kt
@@ -22,8 +22,10 @@
  */
 package com.shopify.checkoutsheetkit
 
+import android.Manifest
 import android.annotation.SuppressLint
 import android.content.Context
+import android.content.pm.PackageManager
 import android.graphics.Color.TRANSPARENT
 import android.os.Build
 import android.util.AttributeSet
@@ -31,6 +33,7 @@ import android.view.KeyEvent
 import android.view.View
 import android.view.ViewGroup.LayoutParams
 import android.view.ViewGroup.LayoutParams.MATCH_PARENT
+import android.webkit.PermissionRequest
 import android.webkit.RenderProcessGoneDetail
 import android.webkit.WebChromeClient
 import android.webkit.WebResourceError
@@ -38,6 +41,8 @@ import android.webkit.WebResourceRequest
 import android.webkit.WebResourceResponse
 import android.webkit.WebView
 import android.webkit.WebViewClient
+import androidx.activity.ComponentActivity
+import androidx.core.app.ActivityCompat
 import java.net.HttpURLConnection.HTTP_GONE
 import java.net.HttpURLConnection.HTTP_NOT_FOUND
 
@@ -64,6 +69,32 @@ internal abstract class BaseWebView(context: Context, attributeSet: AttributeSet
             override fun onProgressChanged(view: WebView?, newProgress: Int) {
                 super.onProgressChanged(view, newProgress)
                 getEventProcessor().updateProgressBar(newProgress)
+            }
+            override fun onPermissionRequest(request: PermissionRequest) {
+                request.resources?.forEach { resource ->
+                    if (resource == PermissionRequest.RESOURCE_VIDEO_CAPTURE) {
+                        if (!hasCameraPermission()) {
+                            requestPermissions()
+                            request.deny()
+                        } else {
+                            request.grant(request.resources)
+                        }
+                    }
+                }
+            }
+
+            private fun hasCameraPermission() =
+                ActivityCompat.checkSelfPermission(
+                    context,
+                    Manifest.permission.CAMERA
+                ) == PackageManager.PERMISSION_GRANTED
+
+            private fun requestPermissions() {
+                ActivityCompat.requestPermissions(
+                    context as ComponentActivity,
+                    arrayOf(Manifest.permission.CAMERA),
+                    CAMERA_PERMISSION_REQUEST
+                )
             }
         }
         isHorizontalScrollBarEnabled = false
@@ -194,6 +225,7 @@ internal abstract class BaseWebView(context: Context, attributeSet: AttributeSet
     companion object {
         private const val DEPRECATED_REASON_HEADER = "X-Shopify-API-Deprecated-Reason"
         private const val LIQUID_NOT_SUPPORTED = "checkout_liquid_not_supported"
+        private const val CAMERA_PERMISSION_REQUEST = 1
 
         private const val TOO_MANY_REQUESTS = 429
         private val CLIENT_ERROR = 400..499